refactor: publisher page
