+++ apps/journeys-admin/pages/publisher/index.tsx
@@ -9,10 +9,15 @@ import { NextSeo } from 'next-seo'
 import { ReactElement, useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { Role } from '../../__generated__/globalTypes'
+import {
+  GetAdminJourneys,
+  GetAdminJourneysVariables
+} from '../../__generated__/GetAdminJourneys'
+import { JourneyStatus, Role } from '../../__generated__/globalTypes'
 import { PageWrapper } from '../../src/components/PageWrapper'
 import { TemplateList } from '../../src/components/TemplateList'
 import { initAndAuthApp } from '../../src/libs/initAndAuthApp'
+import { GET_ADMIN_JOURNEYS } from '../../src/libs/useAdminJourneysQuery/useAdminJourneysQuery'
 import { useUserRoleQuery } from '../../src/libs/useUserRoleQuery'
 
 function TemplateIndex(): ReactElement {
@@ -42,11 +47,11 @@ function TemplateIndex(): ReactElement {
 
 export const getServerSideProps = withUserTokenSSR({
   whenUnauthed: AuthAction.REDIRECT_TO_LOGIN
-})(async ({ user, locale, resolvedUrl }) => {
+})(async ({ user, locale, resolvedUrl, query }) => {
   if (user == null)
     return { redirect: { permanent: false, destination: '/users/sign-in' } }
 
-  const { redirect, translations } = await initAndAuthApp({
+  const { apolloClient, redirect, translations } = await initAndAuthApp({
     user,
     locale,
     resolvedUrl
@@ -54,8 +59,40 @@ export const getServerSideProps = withUserTokenSSR({
 
   if (redirect != null) return { redirect }
 
+  let variables: GetAdminJourneysVariables = {}
+
+  switch (query.tab ?? 'active') {
+    case 'active':
+      variables = {
+        // from src/components/TemplateList/ActiveTemplateList useAdminJourneysQuery
+        status: [JourneyStatus.draft, JourneyStatus.published],
+        template: true
+      }
+      break
+    case 'archived':
+      variables = {
+        // from src/components/TemplateList/ArchivedTemplateList useAdminJourneysQuery
+        status: [JourneyStatus.archived],
+        template: true
+      }
+      break
+    case 'trashed':
+      variables = {
+        // from src/components/TemplateList/TrashedTemplateList useAdminJourneysQuery
+        status: [JourneyStatus.trashed],
+        template: true
+      }
+      break
+  }
+
+  await apolloClient.query<GetAdminJourneys, GetAdminJourneysVariables>({
+    query: GET_ADMIN_JOURNEYS,
+    variables
+  })
+
   return {
     props: {
+      initialApolloState: apolloClient.cache.extract(),
       ...translations
     }
   }
