--- apps/journeys-admin/src/components/TemplateCard/index.ts
@@ -1,1 +1,0 @@
-export { TemplateCard } from './TemplateCard'
