--- apps/journeys-admin/src/components/TemplateLibrary/index.ts
@@ -1,1 +1,0 @@
-export { TemplateLibrary } from './TemplateLibrary'
