+++ apps/journeys-admin/src/components/TemplateList/ActiveTemplateList/ActiveTemplateList.tsx
@@ -3,11 +3,10 @@ import Box from '@mui/material/Box'
 import Card from '@mui/material/Card'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
+import dynamic from 'next/dynamic'
 import { useSnackbar } from 'notistack'
 import { ReactElement, useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-
-import { Dialog } from '@core/shared/ui/Dialog'
 
 import { JourneyStatus } from '../../../../__generated__/globalTypes'
 import { useAdminJourneysQuery } from '../../../libs/useAdminJourneysQuery'
@@ -15,11 +14,21 @@ import {
   ARCHIVE_ACTIVE_JOURNEYS,
   TRASH_ACTIVE_JOURNEYS
 } from '../../JourneyList/ActiveJourneyList/ActiveJourneyList'
-import { JourneyListProps } from '../../JourneyList/JourneyList'
+import type { JourneyListProps } from '../../JourneyList/JourneyList'
 import { sortJourneys } from '../../JourneyList/JourneySort/utils/sortJourneys'
-import { TemplateCard } from '../../TemplateCard'
+import { LoadingTemplateList } from '../LoadingTemplateList'
+import { TemplateListItem } from '../TemplateListItem'
 
-export function ActiveTemplates({
+const Dialog = dynamic(
+  async () =>
+    await import(
+      /* webpackChunkName: "core/shared/ui-dynamic/Dialog" */
+      '@core/shared/ui-dynamic/Dialog'
+    ).then((mod) => mod.Dialog),
+  { ssr: false }
+)
+
+export function ActiveTemplateList({
   sortOrder,
   event
 }: Omit<JourneyListProps, 'user'>): ReactElement {
@@ -49,8 +58,10 @@ export function ActiveTemplates({
       }
     }
   })
-  const [archiveDialogOpen, setArchiveDialogOpen] = useState(false)
-  const [trashDialogOpen, setTrashDialogOpen] = useState(false)
+  const [archiveDialogOpen, setArchiveDialogOpen] = useState<
+    boolean | undefined
+  >()
+  const [trashDialogOpen, setTrashDialogOpen] = useState<boolean | undefined>()
 
   async function handleArchiveSubmit(): Promise<void> {
     try {
@@ -107,45 +118,38 @@ export function ActiveTemplates({
 
   return (
     <>
-      <Box>
-        {sortedJourneys != null ? (
-          <>
-            {sortedJourneys.map((journey) => (
-              <TemplateCard
-                key={journey.id}
-                journey={journey}
-                isPublisher
-                refetch={refetch}
-              />
-            ))}
-            {sortedJourneys.length === 0 && (
-              <Card
-                variant="outlined"
-                sx={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  pt: 20,
-                  pb: 16,
-                  borderBottomLeftRadius: { xs: 0, sm: 12 },
-                  borderBottomRightRadius: { xs: 0, sm: 12 },
-                  borderTopLeftRadius: 0,
-                  borderTopRightRadius: 0
-                }}
-              >
-                <Typography variant="subtitle1" align="center" gutterBottom>
-                  {t('No templates to display.')}
-                </Typography>
-              </Card>
-            )}
-          </>
-        ) : (
-          <>
-            {[0, 1, 2].map((index) => (
-              <TemplateCard key={`templateCard${index}`} isPublisher />
-            ))}
-          </>
-        )}
-      </Box>
+      {sortedJourneys != null ? (
+        <Box>
+          {sortedJourneys.map((journey) => (
+            <TemplateListItem
+              key={journey.id}
+              journey={journey}
+              refetch={refetch}
+            />
+          ))}
+          {sortedJourneys.length === 0 && (
+            <Card
+              variant="outlined"
+              sx={{
+                display: 'flex',
+                flexDirection: 'column',
+                pt: 20,
+                pb: 16,
+                borderBottomLeftRadius: { xs: 0, sm: 12 },
+                borderBottomRightRadius: { xs: 0, sm: 12 },
+                borderTopLeftRadius: 0,
+                borderTopRightRadius: 0
+              }}
+            >
+              <Typography variant="subtitle1" align="center" gutterBottom>
+                {t('No templates to display.')}
+              </Typography>
+            </Card>
+          )}
+        </Box>
+      ) : (
+        <LoadingTemplateList hideHelperText />
+      )}
       <Stack alignItems="center">
         <Typography
           variant="caption"
@@ -158,44 +162,48 @@ export function ActiveTemplates({
           )}
         </Typography>
       </Stack>
-      <Dialog
-        open={archiveDialogOpen}
-        onClose={handleClose}
-        dialogTitle={{
-          title: 'Archive Templates',
-          closeButton: true
-        }}
-        dialogAction={{
-          onSubmit: handleArchiveSubmit,
-          submitLabel: 'Archive',
-          closeLabel: 'Cancel'
-        }}
-      >
-        <Typography>
-          {t(
-            'Are you sure you would like to archive all active templates immediately?'
-          )}
-        </Typography>
-      </Dialog>
-      <Dialog
-        open={trashDialogOpen}
-        onClose={handleClose}
-        dialogTitle={{
-          title: t('Trash Templates'),
-          closeButton: true
-        }}
-        dialogAction={{
-          onSubmit: handleTrashSubmit,
-          submitLabel: t('Trash'),
-          closeLabel: t('Cancel')
-        }}
-      >
-        <Typography>
-          {t(
-            'Are you sure you would like to trash all active templates immediately?'
-          )}
-        </Typography>
-      </Dialog>
+      {archiveDialogOpen != null && (
+        <Dialog
+          open={archiveDialogOpen}
+          onClose={handleClose}
+          dialogTitle={{
+            title: 'Archive Templates',
+            closeButton: true
+          }}
+          dialogAction={{
+            onSubmit: handleArchiveSubmit,
+            submitLabel: 'Archive',
+            closeLabel: 'Cancel'
+          }}
+        >
+          <Typography>
+            {t(
+              'Are you sure you would like to archive all active templates immediately?'
+            )}
+          </Typography>
+        </Dialog>
+      )}
+      {trashDialogOpen != null && (
+        <Dialog
+          open={trashDialogOpen}
+          onClose={handleClose}
+          dialogTitle={{
+            title: t('Trash Templates'),
+            closeButton: true
+          }}
+          dialogAction={{
+            onSubmit: handleTrashSubmit,
+            submitLabel: t('Trash'),
+            closeLabel: t('Cancel')
+          }}
+        >
+          <Typography>
+            {t(
+              'Are you sure you would like to trash all active templates immediately?'
+            )}
+          </Typography>
+        </Dialog>
+      )}
     </>
   )
 }
