+++ apps/journeys-admin/src/components/TemplateList/ActiveTemplateList/index.ts
@@ -1,0 +1,1 @@
+export { ActiveTemplateList } from './ActiveTemplateList'
