--- apps/journeys-admin/src/components/TemplateList/ActiveTemplates/index.ts
@@ -1,1 +1,0 @@
-export { ActiveTemplates } from './ActiveTemplates'
