+++ apps/journeys-admin/src/components/TemplateList/ArchivedTemplateList/index.ts
@@ -1,0 +1,1 @@
+export { ArchivedTemplateList } from './ArchivedTemplateList'
