--- apps/journeys-admin/src/components/TemplateList/ArchivedTemplates/index.ts
@@ -1,1 +1,0 @@
-export { ArchivedTemplates } from './ArchivedTemplates'
