+++ apps/journeys-admin/src/components/TemplateList/LoadingTemplateList/index.ts
@@ -1,0 +1,1 @@
+export { LoadingTemplateList } from './LoadingTemplateList'
