+++ apps/journeys-admin/src/components/TemplateList/TemplateList.tsx
@@ -1,14 +1,43 @@
 import Box from '@mui/material/Box'
 import Container from '@mui/material/Container'
+import dynamic from 'next/dynamic'
 import { ReactElement, useState } from 'react'
 
-import { JourneyListEvent, JourneyListProps } from '../JourneyList/JourneyList'
+import type {
+  JourneyListEvent,
+  JourneyListProps
+} from '../JourneyList/JourneyList'
 import { SortOrder } from '../JourneyList/JourneySort'
 import { StatusTabPanel } from '../StatusTabPanel'
 
-import { ActiveTemplates } from './ActiveTemplates'
-import { ArchivedTemplates } from './ArchivedTemplates'
-import { TrashedTemplates } from './TrashedTemplates'
+import { LoadingTemplateList } from './LoadingTemplateList'
+
+const ActiveTemplateList = dynamic(
+  async () =>
+    await import(
+      /* webpackChunkName: "ActiveTemplateList" */
+      './ActiveTemplateList'
+    ).then((mod) => mod.ActiveTemplateList),
+  { loading: () => <LoadingTemplateList /> }
+)
+
+const ArchivedTemplateList = dynamic(
+  async () =>
+    await import(
+      /* webpackChunkName: "ArchivedTemplates" */
+      './ArchivedTemplateList'
+    ).then((mod) => mod.ArchivedTemplateList),
+  { loading: () => <LoadingTemplateList /> }
+)
+
+const TrashedTemplateList = dynamic(
+  async () =>
+    await import(
+      /* webpackChunkName: "TrashedTemplateList" */
+      './TrashedTemplateList'
+    ).then((mod) => mod.TrashedTemplateList),
+  { loading: () => <LoadingTemplateList /> }
+)
 
 export function TemplateList(): ReactElement {
   const [sortOrder, setSortOrder] = useState<SortOrder>()
@@ -31,9 +60,9 @@ export function TemplateList(): ReactElement {
     <Box sx={{ mx: { xs: -6, sm: 0 } }} data-testid="JourneysAdminTemplateList">
       <Container disableGutters>
         <StatusTabPanel
-          activeList={<ActiveTemplates {...journeyListProps} />}
-          archivedList={<ArchivedTemplates {...journeyListProps} />}
-          trashedList={<TrashedTemplates {...journeyListProps} />}
+          activeList={<ActiveTemplateList {...journeyListProps} />}
+          archivedList={<ArchivedTemplateList {...journeyListProps} />}
+          trashedList={<TrashedTemplateList {...journeyListProps} />}
           setActiveEvent={handleClick}
           setSortOrder={setSortOrder}
           sortOrder={sortOrder}
