+++ apps/journeys-admin/src/components/TemplateList/TemplateListItem/index.ts
@@ -1,0 +1,1 @@
+export { TemplateListItem } from './TemplateListItem'
