+++ apps/journeys-admin/src/components/TemplateList/TrashedTemplateList/index.ts
@@ -1,0 +1,1 @@
+export { TrashedTemplateList } from './TrashedTemplateList'
