--- apps/journeys-admin/src/components/TemplateList/TrashedTemplates/index.ts
@@ -1,1 +1,0 @@
-export { TrashedTemplates } from './TrashedTemplates'
